+++ render/Cargo.toml
@@ -18,15 +18,14 @@ features = ["cache", "dump"]
 git = "https://github.com/pdf-rs/pdf"
 
 [dependencies]
-pathfinder_renderer = { git = "https://github.com/servo/pathfinder" }
-pathfinder_color = { git = "https://github.com/servo/pathfinder" }
-pathfinder_geometry = { git = "https://github.com/servo/pathfinder" }
-pathfinder_resources = { git = "https://github.com/servo/pathfinder" }
-pathfinder_content = { git = "https://github.com/servo/pathfinder" }
+# pathfinder_renderer = { git = "https://github.com/servo/pathfinder" }
+# pathfinder_color = { git = "https://github.com/servo/pathfinder" }
+# pathfinder_geometry = { git = "https://github.com/servo/pathfinder" }
+# pathfinder_resources = { git = "https://github.com/servo/pathfinder" }
+# pathfinder_content = { git = "https://github.com/servo/pathfinder" }
 log = "0.4"
 font = { git = "https://github.com/pdf-rs/font", features = ["cff", "truetype"] }
-#font = { path = "../../font", features = ["cff", "opentype"] }
-pdf_encoding = "0.4"
+pdf_encoding = {git="https://github.com/pdf-rs/encoding"}
 itertools = "*"
 image = "0.24.2"
 instant = "*"
@@ -37,6 +36,7 @@ once_cell = "*"
 serde_json = "*"
 glyphmatcher = { git = "https://github.com/s3bk/glyphmatcher", optional = true }
 vello_encoding = "*"
+vello = "0.2.0"
 
 [dev-dependencies]
 criterion = "0.3"
