+++ render/src/backend.rs
@@ -1,4 +1,4 @@
-use vello::kurbo::{Affine, Vec2 as Vector2F, Rect as RectF};
+use vello::kurbo::{Affine, Rect as RectF, Stroke};
 
 use pathfinder_content::{
     outline::Outline,
@@ -29,9 +29,12 @@ pub trait Backend {
     fn set_view_box(&mut self, r: RectF);
     fn draw_image(&mut self, xref: Ref<XObject>, im: &ImageXObject, resources: &Resources, transform: Affine, mode: BlendMode, clip: Option<Self::ClipPathId>, resolve: &impl Resolve);
     fn draw_inline_image(&mut self, im: &Arc<ImageXObject>, resources: &Resources, transform: Affine, mode: BlendMode, clip: Option<Self::ClipPathId>, resolve: &impl Resolve);
-    fn draw_glyph(&mut self, font: &FontRc<Self::Encoder>, glyph: &Glyph<Self::Encoder>, mode: &DrawMode, transform: Affine, clip: Option<Self::ClipPathId>);
-    fn get_font(&mut self, font_ref: &MaybeRef<PdfFont>, resolve: &impl Resolve) -> Result<Option<Arc<FontEntry<Self::Encoder>>>, PdfError>;
-    fn add_text(&mut self, span: TextSpan<Self::Encoder>, clip: Option<Self::ClipPathId>);
+    fn draw_glyph(&mut self, glyph: &Glyph, mode: &DrawMode, transform: Affine, clip: Option<Self::ClipPathId>) {
+        self.draw(&glyph.path, mode, FillRule::EvenOdd, transform, clip);
+    }
+    fn get_font(&mut self, font_ref: &MaybeRef<PdfFont>, resolve: &impl Resolve) -> Result<Option<Arc<FontEntry>>, PdfError>;
+    fn add_text(&mut self, span: TextSpan, clip: Option<Self::ClipPathId>);
+
     /// The following functions are for debugging PDF files and not relevant for rendering them.
     fn bug_text_no_font(&mut self, data: &[u8]) {}
     fn bug_text_invisible(&mut self, text: &str) {}
@@ -48,9 +51,10 @@ pub struct FillMode {
 }
 pub enum DrawMode {
     Fill { fill: FillMode },
-    Stroke { stroke: FillMode, stroke_mode: Stroke },
-    FillStroke { fill: FillMode, stroke: FillMode, stroke_mode: Stroke },
+    Stroke { fillMode: FillMode, stoke: Stroke },
+    FillStroke { fill: FillMode, fillMode: FillMode, stoke: Stroke },
 }
+
 impl DrawMode {
     pub fn fill(&self) -> Option<&FillMode> {
         match self {
@@ -60,15 +64,8 @@ impl DrawMode {
     }
     pub fn stroke(&self) -> Option<(&FillMode, &Stroke)> {
         match self {
-            DrawMode::FillStroke { stroke, stroke_mode, .. } | DrawMode::Stroke { stroke, stroke_mode } => Some((stroke, stroke_mode)),
+            DrawMode::FillStroke { fillMode, stoke, .. } | DrawMode::Stroke { fillMode, stoke } => Some((fillMode, stoke)),
             _ => None
         }
     }
-}
-
-
-#[derive(Clone, Debug)]
-pub struct Stroke {
-    pub dash_pattern: Option<(Vec<f32>, f32)>,
-    pub style: StrokeStyle,
 }
